Move memory mocking utils to a separate file
