+++ test/CMakeLists.txt
@@ -14,7 +14,7 @@ find_library(MATH_LIBRARY m)
 foreach (TEST ${TESTS})
     string(REGEX REPLACE ".*/([^/]+).c" "\\1" NAME ${TEST})
     message("Adding test ${NAME}")
-    add_executable(${NAME} "${NAME}.c" assertions.c stream_expectations.c)
+    add_executable(${NAME} "${NAME}.c" assertions.c stream_expectations.c test_allocator.c)
     target_link_libraries(${NAME} ${CMOCKA_LIBRARIES})
     target_link_libraries(${NAME} cbor)
     if(MATH_LIBRARY)
